Polished transmutation mechanics.
The altar only places items in a hopper if that hopper is not indirectly
powered.
If a handler returns one item stack, the altar's target slot is empty,
and the item cannot be placed in a hopper below the altar, the target
slot is changed to the returned item stack.
E.x. if an Echir Sword can is smelted, it remains in the altar to be
transmuted.