Improved central ruins generation.
Now, it generates in a hollowed-out dome rather than a cylinder, which
looks neater and allows for nifty overhangs.
It now includes a crafting bench below the crying obsidian platform.
Increased the chance of diamond armor spawns in the chest, especially
chestplates and leggings (removed helmets). In turn, decreased the rate
of berry spawns.
The enchantment applied to looted diamond armor in the central ruins is
now unbreaking 1, and sometimes protection 1.
Decreased the chance of a smite sword being lootable.
The ruins can now generate underground! Yes, super-safe underground
spawns are back!