Implemented alt nourishment.
Decay is NYI and this is going to need a LOT of balance work.
Most normal food is inedible. Nourishment is obtained from nourishment orbs, which generally drop from Nyxian enemies and from bloodstone crystals.
Sanguine obsidian provides a very strong form of health and food regeneration instead of weaker regeneration and resistance.