Tower changes and worldgen bugfixes.
The towers generally appear more physically intact.
Also fixed a bug with pale brick that was resulting in problematic worldgen.
I mean, this is a really funny bug because the exact same mistake was made at one point with Nyxian stone, and it was causing huge sections of the world to generate as gravel.
If I didn't know any better, I'd say two different people were coding this dumpster fire.