Implemented a basic climbing system in Nyx.
Permafrost can be climbed as if ladders were on it.
The climbing system is subject to a few quirks.
For one, it breaks down when mixed with any of Nyx's other airlike blocks, notably at high altitudes. I might add in a system whereby this block can temporarily take on properties of the other airlike blocks in exchange for being able to repalce them.
Additionally, it's somewhat awkward to descend using it.
Also, isCollidedHorizontally is EVIL.