Transfusion altar changes.
First, the name. This is going to take some getting used to.
Second, there are now broken variants of the transfusion altar. Nine of them can be found in the central ruins, and they are additionally what drops when mining a complete altar.
Broken transfusion altars must be repaired using 8 cortra (doesn't matter if it's dust or crystal). A partially repaired altar or a fully repaired altar will drop some of its cortra.
Added a new material called Sanctified Obsidian. It's made by transfusing alabaster into obsidian.
Transfusion tables are now made with one draconium, four cortra crystals, and one sanctified obsidian. Yes, it is indeed a right pain in the ass to make them now.