Doing something possibly bad.
provideChunk() *seems* to never be called from a multithreaded environment.
Assuming that's the case (and initial tests suggest I'm right), the block and byte arrays it uses have been preallocated, and are reused for each call of provideChunk().
I can imagine that certain optimization mods would cause problems with this, but it's not like this class was thread safe to begin with.
In other news, there's some other minor bugfixes, tweaks, and increased threading.