Skeleton eyesight buffs.
Seeing as the changed "in front of" system was an abject failure and I'm getting conflicting information on what rotationYaw actually is and am not one for looking through the code to confirm, reverted the system to what doesn't result in obscenely low aggro.
The skeletons now arc shots with signficantly higher accuracy.