Bugfixes and attention to detail.
The block blocking code is no longer laughably incorrect. Them's the typos.
Entities that spawn in Nyx burning are extinguished. This means fire arrows.
Entities are extinguished upon taking fire damage in Nyx. The practical upshot of this is that nothing burns on Nyx for particularly long. It's a somewhat limiting and dirty solution, but it sort of works.