Most potions can now be transfused into saline flasks.
Restrictions:
Instant effects need to be level 2, and others need to be extended.
Only positive saline potions are implemented, plus harming, minus saturation.
I really need to a better interface for transfusion handlers. The current system is fine for when there's only one possible transfusion, but the number of cases where that still holds true is converging on zero. Otherwise, there need to be rechecks after rechecks.