Worldgen overhaul.
2018 welcomes and hates you. Or at least me.
Created a brand new biome selection system. At the moment, all it does is checkerboard the world with mountains and forested hills, but in the future it should allow for less guesswork on my part about where biomes are generated.
Did a fair amount of biome tweaking that isn't visible at the moment. Changes include:
* Infested biomes now having infested trees again.
* Rugged biome > exousic biome, complete with exousic ore and a thin layer of exousia on many blocks. Incentive to actually brave that deadly hellscape coming later.
* Tweaked the biome colors, which may or may not be wrong and may or may not be irrelevant. Further testing shall reveal the truth.
* Necromancers are now a higher level spawn and exist outside of a certain biome.
Reduced the ice layer for exousic water. It's now generated only over kinda shallow exousic water. It also now only regenerates over extremely shallow exousic water. The algorithm for this is a bit cheap, so it can also generate iceberg-like things.
Towers now generate a shaped stone pillar to stand on when over exousic water.
There's a fair bit of new boilerplate, including a centralized executor for most of Nyx's computation (esp. world-gen) needs. There's also a fair amount of new code that's wildly incomplete, but should give some idea as to what the plans are.