Implemented the basics of infested walkers.
For the time being, they're very durable and wield pickaxes. Their primary gimmick at the moment is that they are immune to most knockback and go for the head with their pickaxes, causing helmet damage and inflicting grevious damage to those without a helmet.
Added infested phylacteries around the mines. These require a tier 4 (Navistra) pickaxe to destroy, and gradually spawn infested entities.
Extended the branches of the mines.