Moar crafting-related transfusion altar upgrades!
The transfusion altar can now "un-craft" items. The catch is that the recipe that creates the target and catalyst (which should be identical) must produce more items than it consumes (so Sticks > Planks works, but Crafting Table > Planks does not). This takes precedence over the other crafting handlers, because at the end of the day crafting benches exist in vanilla, and an uncrafter, however conditional, does not.
Added a cortra "spark" when the altar begins transfusion.