Added a number of things.
Whoops. Clockworks were not meant to be included in an earlier commit. All well. They're added.
Transfusion altars now require a sentient draconium clockwork instead of plain draconium.
The mines now have some shoddy minecart rails in them.
Added rewards to the mines. Note that amber is one of these rewards and is non-present in ore form. This may change, but it's probably not gonna.
Tweaked spawns in various biomes.
Added biome size weights and generally reduced biome sizes across the board.
Snow generates in thicker piles, larger overhangs... and now collapses on walking (in addition to falling upon) and even if the overhang is two blocks of snow thick.