PreAlpha 17.1 inbound. Fuck.
Fixed a couple of critical game-breaking bugs, one of which was long-standing and missed due to testing in creative mode on my end, and the other of which was missed due to not testing clean builds. This is me getting what I deserve.
Added a blacklist for some items that are not meant to be uncrafted.
Added amber nuggets and programmable clockworks, each deconstructs from their larger counterparts.
Amber is now single-use, per suggestion. Rather than outright destroying them, however, using them breaks them into nuggets.
Added the ability to localize some of IaS2's messages. Not that there's ever likely to be a need for it, but it's a good habit.
Added a system for adding information to tooltips that gets revealed on a shift right click. This is used to add descriptive text and hints for now, but will have more critical informative uses come PreAlpha 20.
Exousium tools now require dust instead of rocks to make.