Cortra tool adjustments!
This may eventually require another change as one of the things in the pipeline is a mining crossbow.
Cortra tools, instead of giving bonus experience (may change to be in addition to) instantly put mined items at the player's location for instant pick-up. This allows using cortra to mine in significantly more precarious locations.