More grenade work.
Added the ice wall grenade, which triggers a brief moment after impact.
On a related note, impact grenades work now.
There is now a double explosion for explosive grenades.  One explosion is for the damage, the other is for terrain destruction.
In order to arm a grenade, the user must now be sneaking. This is to provide a simplified way to stop the player from auto-priming stacked grenades after one goes off automatically, as well as to reduce accidental misclicks.
Grenades now have custom localized text to display what kind of grenade they are (though this isn't likely to be as necessary as it would be for potions).
Grenades now make a sound when their fuses are lit.