Long-standing bugfixes and enchancements.
Fixed Nyxian spades not working on non-Nyx shovelable blocks, and vice versa.
Speaking of consistency, renamed many instances of transmutation in the code to transfusion. The main outliers are the transfusion handler interface and the unlocalized name.
Added some localized text for grenades and one bit that was missing for the transfusion altar.