Started work on remotes.
Grab a remote. Push the button. Get nothing. My cookie. Screw you.
Added a second type of grenade: remote grenades. When holding a grenade detonator and using it, any thrown remote grenades will have their fuses lit on a greatly reduced timer (at most half of the normal timer). Remote grenades will spuriously detonate if left alive for too long.
Removed the ability for certain grenades to self-trigger on impact.
Honestly the code for all this is a little bit horrifying, but it has no immediate and obvious bugs, so...